Refactor provider registration and config pipeline
Streamline and modularize provider and policy registration for OpenFeature DI integration. Remove legacy interfaces and builder code, consolidate extension methods, and introduce a clear separation between build-time configuration and runtime options. Ensure provider options are registered/configured once, and improve maintainability by encapsulating registration logic in dedicated components.
