+++ src/OpenFeature.Providers.DependencyInjection/Configuration/OpenFeatureProviderConfigurationExtensions.cs
@@ -18,4 +18,21 @@ public static class OpenFeatureProviderConfigurationExtensions
 
         return state.GetOrAddState(static () => new OpenFeatureProviderConfiguration());
     }
+
+    /// <summary>
+    /// Applies provider configuration to options.
+    /// </summary>
+    /// <param name="configuration">The provider configuration.</param>
+    /// <param name="options">The options to populate.</param>
+    public static void ApplyTo(this OpenFeatureProviderConfiguration configuration, OpenFeatureProviderOptions options)
+    {
+        if (configuration is null) throw new ArgumentNullException(nameof(configuration));
+        if (options is null) throw new ArgumentNullException(nameof(options));
+
+        options.HasDefaultProvider = configuration.HasDefaultProvider;
+        foreach (var domain in configuration.Domains)
+        {
+            options.AddProviderName(domain);
+        }
+    }
 }
